server/webserver/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
SQLAlchemy==2.0.32
pymysql==1.1.1
pydantic==2.8.2
cryptography==42.0.8
paho-mqtt==2.1.0